+++ src/types.ts
@@ -1,0 +1,4 @@
+import { ProblemDetails } from "@tectonique/api-standards";
+
+export type FallbackInternalServerProblemDetailGenerator =
+  () => ProblemDetails.ProblemDetail<any, any, any>;