+++ CHANGELOG.md
@@ -13,3 +13,5 @@
 
 ## 0.0.1
 - Initial release.
+- Envelope interceptor.
+- ProblemDetail exception filter.